refactor(T11832): delete unreachable OpenAITransport — collapse 7→6 transport classes
OpenAITransport (transports/openai.ts, 880 LOC) was a fully-implemented but
UNREACHABLE transport — 0 live construction sites, superseded on the live
ModelRunner path by ChatCompletionsTransport (the generic OpenAI-compatible
transport that already handles every openai-compat provider + gemini/kimi quirks
via its provider-quirk dispatch table + the ProviderProfile hooks).

This is the concrete "collapse the 7 transport classes" reduction for the one
genuinely-removable class (T11767 follow-up). The remaining 6 (anthropic_messages,
chat_completions, codex_responses, ollama_native NDJSON, bedrock_converse, gemini)
are distinct LIVE wire protocols — correctly one class each, dispatched by the
T11767 Record<ApiMode,TransportCtor> table.

- Deleted transports/openai.ts + its 3 OpenAITransport-only test surfaces
  (migration-snapshot-openai.test.ts, transports/openai.test.ts, the OpenAITransport
  blocks in transports.test.ts + streaming-tool-calls.test.ts).
- Relocated the tested `usesMaxCompletionTokens` model-id predicate to
  model-metadata.ts (its natural home), re-exported from the llm barrel.
- Removed OpenAITransport/OpenAITransportOptions from the transports + llm barrels.

Net -1504 LOC. tsc -b clean; 945 core LLM tests pass.
